chore: move `scrollbar.dart` into /layout
Didn't really make sense for it to be in /indicators
