+++ lib/src/layout/resizable_pane.dart
@@ -1,7 +1,7 @@
 import 'dart:math' as math show max, min;
 
 import 'package:flutter/services.dart' show SystemMouseCursor;
-import 'package:macos_ui/src/indicators/scrollbar.dart';
+import 'package:macos_ui/src/layout/scrollbar.dart';
 import 'package:macos_ui/src/library.dart';
 import 'package:macos_ui/src/theme/macos_theme.dart';
